Created Validate Session Component, Re-structured Contacts JSON
